# apps/backend/src/api/routes/federation.controller.ts
import {
  Body,
  Controller,
  Get,
  Headers,
  HttpException,
  HttpStatus,
  Post,
  Query,
} from '@nestjs/common';
import { ApiOperation, ApiTags } from '@nestjs/swagger';
import { Throttle } from '@nestjs/throttler';
import {
  FEDERATION_AUDIENCE,
  FederationService,
} from '@validpost/nestjs-libraries/database/prisma/federation/federation.service';
import { GetUserFromRequest } from '@validpost/nestjs-libraries/user/user.from.request';
import { GetOrgFromRequest } from '@validpost/nestjs-libraries/user/org.from.request';
import { User, Organization } from '@prisma/client';
import {
  AuthorizeFederationQueryDto,
  ApproveFederationDto,
} from '@validpost/nestjs-libraries/dtos/federation/authorize-federation.dto';
import { TokenFederationDto } from '@validpost/nestjs-libraries/dtos/federation/token-federation.dto';

@ApiTags('Federation')
@Controller('/.well-known')
export class FederationDiscoveryController {
  constructor(private _federationService: FederationService) {}

  @Get('/validpost-identity')
  @ApiOperation({
    summary: 'ValidPost ID discovery document (endpoints, JWKS URI, scopes)',
  })
  discovery() {
    return this._federationService.getDiscoveryDocument();
  }
}

@ApiTags('Federation')
@Controller('/federation')
export class FederationController {
  constructor(private _federationService: FederationService) {}

  @Get('/jwks')
  @ApiOperation({ summary: 'Public JWKS used to verify federation id_tokens' })
  jwks() {
    return this._federationService.getJwks();
  }

  @Get('/authorize')
  @ApiOperation({
    summary: 'Validate a federation authorize request (consent screen metadata)',
  })
  authorize(@Query() query: AuthorizeFederationQueryDto) {
    this._federationService.validateAuthorizeRequest(query.redirect_uri);

    return {
      client: {
        name: 'ValidPost Template Store',
        audience: FEDERATION_AUDIENCE,
        redirectUri: query.redirect_uri,
      },
      state: query.state,
    };
  }

  @Post('/token')
  @Throttle({ default: { limit: 20, ttl: 60000 } })
  @ApiOperation({
    summary: 'Exchange a federation authorization code (public client, PKCE only)',
  })
  token(@Body() body: TokenFederationDto) {
    if (body.grant_type !== 'authorization_code') {
      throw new HttpException(
        { error: 'unsupported_grant_type' },
        HttpStatus.BAD_REQUEST
      );
    }

    return this._federationService.exchangeCode(
      body.code,
      body.redirect_uri,
      body.code_verifier
    );
  }

  // GET and POST are separate handlers on purpose: a stacked @Get+@Post on one
  // method is inert — decorators apply bottom-up and the second overwrites
  // METHOD_METADATA — so it would read as supported while 404ing. OIDC Core
  // 5.3 requires the UserInfo endpoint to support both methods.
  @Get('/userinfo')
  @ApiOperation({ summary: 'Scope-gated identity claims for a posf_ access token' })
  async userinfoGet(@Headers('authorization') authorization?: string) {
    return this.userinfo(authorization);
  }

  @Post('/userinfo')
  @ApiOperation({ summary: 'Scope-gated identity claims for a posf_ access token' })
  async userinfoPost(@Headers('authorization') authorization?: string) {
    return this.userinfo(authorization);
  }

  private async userinfo(authorization?: string) {
    const token = authorization?.startsWith('Bearer ')
      ? authorization.slice('Bearer '.length)
      : undefined;
    if (!token || !token.startsWith('posf_')) {
      throw new HttpException(
        { error: 'invalid_token' },
        HttpStatus.UNAUTHORIZED
      );
    }

    const info = await this._federationService.getUserInfo(token);
    if (!info) {
      throw new HttpException(
        { error: 'invalid_token' },
        HttpStatus.UNAUTHORIZED
      );
    }
    return info;
  }
}

@ApiTags('Federation')
@Controller('/federation')
export class FederationAuthorizedController {
  constructor(private _federationService: FederationService) {}

  @Post('/authorize')
  @ApiOperation({ summary: 'Approve or deny a federation consent request' })
  async approveOrDeny(
    @Body() body: ApproveFederationDto,
    @GetUserFromRequest() user: User,
    @GetOrgFromRequest() org: Organization
  ) {
    this._federationService.validateAuthorizeRequest(body.redirect_uri);

    const redirectUrl = new URL(body.redirect_uri);
    if (body.action === 'deny') {
      redirectUrl.searchParams.set('error', 'access_denied');
      if (body.state) {
        redirectUrl.searchParams.set('state', body.state);
      }
      return { redirect: redirectUrl.toString() };
    }

    const code = await this._federationService.createAuthorizationCode(
      user.id,
      org.id,
      {
        redirectUri: body.redirect_uri,
        codeChallenge: body.code_challenge,
        codeChallengeMethod: body.code_challenge_method,
        nonce: body.nonce,
        scope: body.scope,
      }
    );

    redirectUrl.searchParams.set('code', code);
    if (body.state) {
      redirectUrl.searchParams.set('state', body.state);
    }
    return { redirect: redirectUrl.toString() };
  }
}