# openapi.yml
...
openapi: 3.0.0
paths:
...
  /federation/authorize:
    get:
      operationId: FederationController_authorize
      parameters:
        - name: response_type
          required: true
          in: query
          schema:
            type: string
            enum:
              - code
        - name: redirect_uri
          required: true
          in: query
          schema:
            format: uri
            type: string
        - name: state
          required: false
          in: query
          schema:
            type: string
        - name: nonce
          required: false
          in: query
          schema:
            type: string
        - name: code_challenge
          required: false
          in: query
          schema:
            type: string
        - name: code_challenge_method
          required: false
          in: query
          schema:
            type: string
            enum:
              - S256
        - name: scope
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      summary: Validate a federation authorize request (consent screen metadata)
      tags:
        - Federation
    post:
      operationId: FederationAuthorizedController_approveOrDeny
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveFederationDto"
      responses:
        "201":
          description: ""
      summary: Approve or deny a federation consent request
      tags:
        - Federation
  /federation/token:
    post:
      operationId: FederationController_token
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenFederationDto"
      responses:
        "201":
          description: ""
      summary: Exchange a federation authorization code (public client, PKCE only)
      tags:
        - Federation
  /federation/userinfo:
    get:
      operationId: FederationController_userinfoGet
      parameters:
        - name: authorization
          required: true
          in: header
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      summary: Scope-gated identity claims for a posf_ access token
      tags:
        - Federation
    post:
      operationId: FederationController_userinfoPost
      parameters:
        - name: authorization
          required: true
          in: header
          schema:
            type: string
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
      summary: Scope-gated identity claims for a posf_ access token
      tags:
        - Federation
...
  /ai/media:
    post:
      operationId: AiUserController_createMediaJob
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaJobDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
      tags:
        - AI User
...
  /settings/ai/defaults/{category}:
    put:
      operationId: OrgAiSettingsController_setModelDefault
      parameters:
        - name: category
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetDefaultModelDto"
      responses:
        "200":
          description: ""
      tags:
        - Org AI Settings
    delete:
      operationId: OrgAiSettingsController_clearModelDefault
      parameters:
        - name: category
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Org AI Settings
  /settings/ai/defaults/catalog:
    get:
      operationId: OrgAiSettingsController_getModelDefaultsCatalog
      parameters:
        - name: category
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      tags:
        - Org AI Settings
...
  /settings/media/config/{identifier}:
    put:
      operationId: MediaProviderController_upsertConfig
      parameters:
        - name: identifier
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertMediaConfigDto"
      responses:
        "200":
          description: ""
      tags:
        - Org Media Provider Settings
    delete:
      operationId: MediaProviderController_deleteConfig
      parameters:
        - name: identifier
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Org Media Provider Settings
  /settings/media/config/{identifier}/storage:
    put:
      operationId: MediaProviderController_setStorage
      parameters:
        - name: identifier
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetMediaStorageDto"
      responses:
        "200":
          description: ""
      tags:
        - Org Media Provider Settings
  /settings/media/config/{identifier}/set-active:
    post:
      operationId: MediaProviderController_setActive
      parameters:
        - name: identifier
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetActiveVersionDto"
      responses:
        "201":
          description: ""
      tags:
        - Org Media Provider Settings
  /settings/media/config/{identifier}/test:
    post:
      operationId: MediaProviderController_testConnection
      parameters:
        - name: identifier
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderTestConnectionDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
      tags:
        - Org Media Provider Settings
...
  /settings/content/media-defaults/{category}:
    put:
      operationId: MediaDefaultsController_setMediaDefault
      parameters:
        - name: category
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetDefaultModelDto"
      responses:
        "200":
          description: ""
      tags:
        - Media Defaults
    delete:
      operationId: MediaDefaultsController_clearMediaDefault
      parameters:
        - name: category
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Media Defaults
  /settings/content/media-defaults/catalog:
    get:
      operationId: MediaDefaultsController_getMediaDefaultsCatalog
      parameters:
        - name: category
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      tags:
        - Media Defaults
...
  /brands/{id}/default:
    post:
      operationId: BrandsController_setDefault
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - Brands
...
info:
  title: ValidPost API
  description: OpenAPI description of the ValidPost backend (NestJS). Generated from the controllers — do not hand-edit; run `pnpm run openapi:generate`.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    api-key:
      type: apiKey
      in: header
      name: Authorization
    bearer:
      scheme: bearer
      bearerFormat: Token
      type: http
  schemas:
...
    TokenFederationDto:
      type: object
      properties:
        grant_type:
          type: string
        code:
          type: string
        redirect_uri:
          type: string
          format: uri
        code_verifier:
          type: string
      required:
        - grant_type
        - code
        - redirect_uri
...
    ApproveFederationDto:
      type: object
      properties:
        action:
          type: string
          enum:
            - approve
            - deny
        redirect_uri:
          type: string
          format: uri
        state:
          type: string
        nonce:
          type: string
        code_challenge:
          type: string
        code_challenge_method:
          type: string
          enum:
            - S256
        scope:
          type: string
      required:
        - action
        - redirect_uri
...
    MediaJobDto:
      type: object
      properties: {}
...
    SetActiveVersionDto:
      type: object
      properties: {}
    ProviderTestConnectionDto:
      type: object
      properties: {}
...
    SetDefaultModelDto:
      type: object
      properties:
        providerId:
          type: string
        version:
          type: string
        model:
          type: string
        settings:
          type: object
      required:
        - providerId
...
    UpsertMediaConfigDto:
      type: object
      properties: {}
    SetMediaStorageDto:
      type: object
      properties: {}
...